Resolve Conflicts:
	server/src-lib/Hasura/GraphQL/Schema.hs

+++ console/src/components/Services/EventTrigger/Settings/Settings.js
@@ -1,5 +1,6 @@
 import PropTypes from 'prop-types';
 import React, { Component } from 'react';
+import AceEditor from 'react-ace';
 import TableHeader from '../TableCommon/TableHeader';
 import { deleteTrigger } from '../EventActions';
 
@@ -19,8 +20,12 @@ class Settings extends Component {
     )[0];
     triggerSchema = triggerSchema ? triggerSchema : {};
 
-    const handleDeleteTrigger = () => {
-      dispatch(deleteTrigger(triggerName));
+    const handleDeleteTrigger = e => {
+      e.preventDefault();
+      const isOk = confirm('Are you sure?');
+      if (isOk) {
+        dispatch(deleteTrigger(triggerName));
+      }
     };
 
     return (
@@ -67,7 +72,19 @@ class Settings extends Component {
                 </tr>
                 <tr>
                   <td>Operation / Columns</td>
-                  <td>{JSON.stringify(triggerSchema.definition, null, 4)}</td>
+                  <td>
+                    <AceEditor
+                      mode="json"
+                      theme="github"
+                      name="payload"
+                      value={JSON.stringify(triggerSchema.definition, null, 4)}
+                      minLines={4}
+                      maxLines={100}
+                      width="100%"
+                      showPrintMargin={false}
+                      showGutter={false}
+                    />
+                  </td>
                 </tr>
               </tbody>
             </table>
